backend/requirements.txt
django==1.9.4
django-extensions==1.6.1
django-filter==0.12.0
djangorestframework==3.3.2
psycopg2==2.6.1
django-autofixture==0.12.0
flake8==2.5.4